requirements.txt
pyspark
pandas
psycopg2
python-dotenv
sqlalchemy
psycopg2-binary